careeros-backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.123.3"}
psycopg = {extras = ["binary"], version = "^3.3.0"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"
google-generativeai = "^0.8.5"
pydantic-settings = "^2.12.0"
aiosqlite = "^0.21.0"
libsql-experimental = "^0.0.55"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
